Review changes: - renaming variables - changing assertion methods - renaming file name
