+++ test/integration/Cover/index.js
@@ -1,4 +1,4 @@
 describe('Cover integration tests', function () {
   require('./expireCover');
-  require('./stakingPools');
+  require('./createStakingPool');
 });
